# apps/api/swagger.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - App
  /auth/register:
    post:
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User successfully registered
        '400':
          description: Bad request - validation failed
        '409':
          description: User already exists
      summary: Register a new user
      tags:
        - auth
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Login successful, session cookie set
        '401':
          description: Invalid credentials
      summary: Login user and create session
      tags:
        - auth
  /auth/logout:
    post:
      operationId: AuthController_logout
      parameters: []
      responses:
        '200':
          description: Logout successful
      security:
        - sid: []
      summary: Logout user and destroy session
      tags:
        - auth
  /auth/me:
    get:
      operationId: AuthController_me
      parameters: []
      responses:
        '200':
          description: Returns current user or null
      security:
        - sid: []
      summary: Get current authenticated user
      tags:
        - auth
info:
  title: Group Expense Tracker API
  description: API for managing group expenses and user authentication
  version: '1.0'
  contact: {}
tags:
  - name: auth
    description: Authentication endpoints
servers: []
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: sid
  schemas:
    RegisterDto:
      type: object
      properties:
        email:
          type: string
          description: User email address
          example: user@example.com
        password:
          type: string
          description: User password
          example: SecurePassword123!
        name:
          type: string
          description: User display name
          example: John Doe
      required:
        - email
        - password
    LoginDto:
      type: object
      properties:
        email:
          type: string
          description: User email address
          example: user@example.com
        password:
          type: string
          description: User password
          example: SecurePassword123!
      required:
        - email
        - password